+++ src/typeguard/_memo.py
@@ -2,8 +2,9 @@ from __future__ import annotations
 
 import inspect
 import sys
+from inspect import isclass
 from types import FunctionType
-from typing import TYPE_CHECKING, Any, Mapping
+from typing import TYPE_CHECKING, Any, Dict, ForwardRef, Mapping, Tuple
 from weakref import WeakKeyDictionary
 
 from ._utils import function_name
@@ -36,9 +37,11 @@ class TypeCheckMemo:
 
 
 class CallMemo(TypeCheckMemo):
-    __slots__ = "func", "func_name", "arguments", "type_hints"
+    __slots__ = "func", "func_name", "arguments", "self_type", "type_hints"
 
     arguments: Mapping[str, Any]
+    self: Any
+    self_type: type[Any] | None
 
     def __init__(
         self,
@@ -53,6 +56,12 @@ class CallMemo(TypeCheckMemo):
         self.func_name = function_name(func)
         signature = inspect.signature(func)
 
+        # Assume the first argument is bound as "self"
+        if args:
+            self.self_type = args[0] if isclass(args[0]) else type(args[0])
+        else:
+            self.self_type = None
+
         if args is not None and kwargs is not None:
             self.arguments = signature.bind(*args, **kwargs).arguments
         else:
@@ -64,6 +73,26 @@ class CallMemo(TypeCheckMemo):
         try:
             self.type_hints = _type_hints_map[func]
         except KeyError:
-            self.type_hints = _type_hints_map[func] = get_type_hints(
-                func, localns=frame_locals, include_extras=True
-            )
+            try:
+                self.type_hints = _type_hints_map[func] = get_type_hints(
+                    func, localns=frame_locals, include_extras=True
+                )
+            except NameError:
+                type_hints = {}
+                for key, annotation in func.__annotations__.items():
+                    if type(annotation) is str:
+                        annotation = ForwardRef(annotation)
+
+                    type_hints[key] = annotation
+
+                self.type_hints = type_hints
+
+            for key, annotation in self.type_hints.items():
+                if key == "return":
+                    continue
+
+                param = signature.parameters[key]
+                if param.kind is inspect.Parameter.VAR_POSITIONAL:
+                    self.type_hints[key] = Tuple[annotation, ...]
+                elif param.kind is inspect.Parameter.VAR_KEYWORD:
+                    self.type_hints[key] = Dict[str, annotation]